Remodel new types abstractions a bit
