+++ chimney/src/main/scala-2/io/scalaland/chimney/internal/compiletime/ResultsPlatform.scala
@@ -1,0 +1,7 @@
+package io.scalaland.chimney.internal.compiletime
+
+private[compiletime] trait ResultsPlatform extends Results { this: DefinitionsPlatform =>
+
+  // TODO: proper error generation
+  protected def reportError(errors: String): Nothing = c.abort(c.enclosingPosition, errors)
+}
