+++ chimney/src/main/scala-3/io/scalaland/chimney/compiletime/ResultsPlatform.scala
@@ -1,0 +1,8 @@
+package io.scalaland.chimney.internal.compiletime
+
+private[compiletime] trait ResultsPlatform extends Results { this: DefinitionsPlatform =>
+
+  import quotes.*, quotes.reflect.*
+
+  protected def reportError(errors: String): Nothing = report.errorAndAbort(errors, Position.ofMacroExpansion)
+}
