+++ chimney/src/main/scala/io/scalaland/chimney/internal/compiletime/derivation/Derivation.scala
@@ -1,0 +1,18 @@
+package io.scalaland.chimney.internal.compiletime.derivation
+
+import io.scalaland.chimney.internal.compiletime.Definitions
+import io.scalaland.chimney.{partial, PartialTransformer, Transformer}
+
+import scala.annotation.nowarn
+
+@nowarn("msg=The outer reference in this type test cannot be checked at run time.")
+private[derivation] trait Derivation { this: Definitions & Legacy =>
+
+  protected def instantiateTotalTransformer[From, To](
+      f: Expr[From] => DerivationResult[Expr[To]]
+  ): DerivationResult[Expr[Transformer[From, To]]]
+
+  protected def instantiatePartialTransformer[From: Type, To: Type](
+      f: (Expr[From], Expr[Boolean]) => DerivationResult[Expr[partial.Result[To]]]
+  ): DerivationResult[Expr[PartialTransformer[From, To]]]
+}
